fix: inject event improvements + frontend tests + misc fixes
- Wire target_communities in inject event API + orchestrator
- Add bad_review to allowed event types
- Add frontend cache invalidation on inject success
- Add InjectEventModal tests (14 tests)
- Add EngineControlPanel tests
- Fix propagation animation utils extraction
- Misc linter/formatter fixes across backend and frontend

+++ backend/app/api/schemas.py
@@ -151,7 +151,7 @@ class StepResultResponse(BaseModel):
     adoption_rate: float = 0.0
     mean_sentiment: float = 0.0
     sentiment_variance: float = 0.0
-    diffusion_rate: int = 0
+    diffusion_rate: float = 0.0
     total_adoption: int = 0
     community_metrics: dict[str, Any] = Field(default_factory=dict)
     action_distribution: dict[str, int] = Field(default_factory=dict)
